Restrict friction law include to factory to achieve shorter compile times
